% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankIC.R
\name{rankICpen}
\alias{rankICpen}
\title{Fit the interval-censored AFT model with quantile linear model}
\usage{
rankICpen(
  L,
  R,
  x,
  beta0 = rep(1, ncol(x)),
  type = NULL,
  selection = NULL,
  lamb.len = 50,
  lambmin = NULL,
  r = 3.7,
  outlier = "without",
  outlier.rate = 0.2,
  outlier.value = 5,
  weight = rep(1, length(L)),
  tol = 1e-10
)
}
\arguments{
\item{L}{left-censoring time, having 0 if left-censored.}

\item{R}{right-censoring time, having \code{Inf} if right-censored.}

\item{x}{X matrix of baseline covariates.}

\item{beta0}{true parameter values, including intercept, with the default being 1 for all parameters.}

\item{type}{penalized estimating method, default is "scad", or "lasso", "alasso" penalties are applicable.}

\item{selection}{method to use variable selection method, default is "GCV", or "BIC", "AIC" methods are applicable.}

\item{lamb.len}{the range of the lambda grid, default is 50.}

\item{lambmin}{the minimum value of lambda, default is 2^(-10).}

\item{r}{the value provided in the SCAD penalty. The default value is 3.7.}

\item{outlier}{a logical value indicating whether to replace outlier data with some of the normal data. If set to "TRUE," outliers will be replaced.}

\item{outlier.rate}{the outlier rate to set for the normal data, default is 0.2.}

\item{outlier.value}{The value to add to the normal data in order to generate outlier values, default is 5.}

\item{weight}{weights of covariate x and response variable y. The default is 1.}

\item{tol}{tolerance of estimated as 0 and calculated. The default value is 1e-10.}
}
\value{
\code{rankICpen} returns a data frame containing at least the following components:
\itemize{
\item \code{est}: regression estimator.
\item \code{cor}: the number of correctly estimated non-zero values for \code{est}.
\item \code{incor}: the number of correctly estimated zero values for \code{est}.
\item \code{amad}: calculted AMAD values related abstract bias value for \code{est}.
\item \code{mrme}: calculted MRME values related squared bias value for \code{est}.
\item \code{optmeasure}: optimalized measured values for \code{est}.
\item \code{optlamb}: optimalized lambda values for \code{est}.
\item \code{beta}: optimalized beta value.
}
}
\description{
Fit the interval-censored AFT model with quantile linear model with general interval-censored data
}
\details{
see Kim et al., (2024+) for detailed method explanation.
}
\examples{
\dontrun{
# Simulations
set.seed(111)
n = 200
x1 = runif(n,-1,1)
x2 = rbinom(n,1,0.43)
x = cbind(x1,x2)
T = 4 + x1 + x2 + rnorm(n)
U = 4 + (1 - 0.25*x1)*runif(n, -6, 5)
V = U + (1 - 0.1*x2)*runif(n, 6, 20)
U = exp(dplyr::case_when(TRUE ~ T, T>V ~ V, T<U ~ -Inf))
V = exp(dplyr::case_when(TRUE ~ T, T>V ~ Inf, T<U ~ U))
type="scad"; selection="BIC"; outlier=FALSE
rankICpen(L=U,R=V,x=x,type=type,selection=selection,outlier=outlier)


# Data example
library(PICBayes)
library(tidyverse)
data("mCRC")
d = with(data.frame(mCRC), data.frame(U = ifelse(y==0,R,L),
                                      V = ifelse(y==2,L,R),
                                      # Cluster weighted data
                                      id=(rep(c(table(SITE)),c(table(SITE)))),
                                      # Treatment arm: 0 = FOLFIRI alone, 1 = Panitumumab + FOLFIRI.
                                      x1= case_when(TRT_C == 0 ~ 0, #Pan et al data
                                                    TRT_C == 1 ~ 1),
                                      # Tumor KRAS mutation status: 0 = wild-type, 1 = mutant.
                                      x2= case_when(KRAS_C == 0 ~ 1,
                                                    KRAS_C == 1 ~ 0),
                                      delta = case_when(IC == 0 ~ 1,
                                                        IC == 1 ~ 4)
));
U=(log(d$U));V=log(d$V); type="scad"; selection="BIC"; outlier=FALSE
x = cbind(d$x1,d$x2); id=d$id;  tau=0.1;
rankICpen(L=U,R=V,x=x,type=type,selection=selection,outlier=outlier)
}
}
\references{
Kim, Y., Park, S., Choi, S. (2024+). Rank-based variable selection with interval-censored data.
}
